+++ src/models/user.ts
@@ -1,0 +1,6 @@
+import type { Contact } from "./contact";
+
+export interface User extends Contact {
+  lists?: any;
+  password?: string;
+}