+++ src/types/plugin.ts
@@ -1,27 +1,35 @@
 import ParentStoreClass from "../plugins/parentStore"
 
-import type { Store } from "pinia"
-import type { AnyObject, CustomStore } from "pinia-plugin-subscription"
-import type { ExtendedStoreOptions } from "./store"
+import type { StateTree, Store } from "pinia"
+import type { CustomStore } from "pinia-plugin-subscription"
+import type { ExtendedStoreOptions, ParentStoreOptions } from "./store"
 
+export type ParentStoreResult<TStore extends object, TState extends StateTree> =
+    CustomStore<TStore, TState> | Store
 
-interface ActionFlow {
-    after?: Function | string
-    before?: Function | string
-}
-
-export type ActionFlows = Record<string, ActionFlow>
-
-export interface ParentStoreInterface {
+export interface ParentStoreInterface<
+    TStore extends object = Record<string, never>,
+    TState extends StateTree = StateTree
+> {
     get id(): string
-    build: (childId: string) => CustomStore<AnyObject, AnyObject>
+    get options(): ParentStoreOptions | undefined
+    build: (childId?: string) => ParentStoreResult<TStore, TState>
 }
 
 
-export type ParentStore = (
-    <TStore = AnyObject, TState = AnyObject>(id: string) => (CustomStore<TStore, TState> | Store)
-)
+export type ParentStore<TStore extends object = Record<string, never>, TState extends StateTree = StateTree> =
+    (id: string) => ParentStoreResult<TStore, TState>
 
-export type ParentStoreConstructor = (() => CustomStore<AnyObject, AnyObject> | Store) | ParentStoreClass
+export type ParentStoreConstructor<
+    TStore extends object = Record<string, never>,
+    TState extends StateTree = StateTree
+> = (() => ParentStoreResult<TStore, TState>) | ParentStoreClass<TStore, TState>
 
-export interface PluginStoreOptions extends AnyObject { storeOptions: ExtendedStoreOptions }
+export interface PluginStoreOptions {
+    storeOptions?: Omit<ExtendedStoreOptions, 'propertiesToRename' | 'actionsToRename'> & {
+        /** @Deprecated use ParentStoreOptions instead */
+        actionsToRename?: Record<string, string>
+        /** @Deprecated use ParentStoreOptions instead */
+        propertiesToRename?: Record<string, string>
+    }
+}